Fix minor bugs
